'Quadlit' App
Created a basic drawing canvas that's cross-compatible with all devices
(iOs, blackberry, Android etc.). Also takes advantage of accelerometer
on smart phones to use to draw.

Next step: figure a way to record all touches the user makes, and saves
it some sort of file (ex .txt file) that's sent to a desktop machine,
that makes the calculations for the Quadcopter.